Relocate sqlcmysql
