+++ hack/ccp/internal/store/sqlcmysql/query.sql.go
